Commit juego de adivinar el numero y operador ternario
